requirements.txt
aiogram==3.3.0
pyrogram==2.0.106
tgcrypto==1.2.5
aiofiles==23.2.1
python-dotenv==1.0.0


